build(deps): lock all major versions
Currently, several versions are locked on 0. Because of how semver
works, this means that of these dependencies are allow to freely be
upgraded with breaking changes. Essentially, it's equivalent to setting
*, only as long as it doesn't go to a "one dot" release.

Most of the cargo ecosystem is pre one dot, and will be so forever, so
this is extremely brittle.

faythe v0.1.0 (/home/ces/org/src/dbc/faythe/.worktree/pin-crates)
├── acme-lib v0.8.1 (https://github.com/DBCDK/acme-lib?branch=dbc-fork#6ae27e4e)
├── async-trait v0.1.83 (proc-macro)
├── base64 v0.22.1
├── chrono v0.4.38
├── clap v4.4.18
├── dbc-rust-modules v0.1.0 (https://github.com/dbcdk/rust-modules?branch=main#ab38efd1)
├── env_logger v0.11.5
├── lazy_static v1.5.0
├── num-traits v0.2.19
│   [build-dependencies]
├── openssl v0.10.66
├── prometheus_exporter_base v1.4.0
├── regex v1.11.0
├── reqwest v0.12.7
├── serde v1.0.210
├── serde_derive v1.0.210 (proc-macro)
├── serde_json v1.0.128
├── tokio v1.40.0
├── trust-dns-resolver v0.23.2
├── url v2.5.2
├── vaultrs v0.7.2
├── vaultrs-login v0.2.1
└── walkdir v2.5.0

Signed-off-by: Christina Sørensen <christina@cafkafk.com>