remove dead code warnings as per rustc
Leaves in one:

```
warning: field `0` is never read
   --> src/issuer.rs:246:28
    |
246 |     NoIpsForResolversFound(&'l String),
    |     ---------------------- ^^^^^^^^^^
    |     |
    |     field in this variant
    |
    = note: `#[warn(dead_code)]` on by default
help: consider changing the field to be of unit type to suppress this warning while preserving the field numbering, or remove the field
    |
246 |     NoIpsForResolversFound(()),
    |
```

That one should probably be dealt with after the next nixpkgs bump,
because it'll start affecting every field that is only read via derived
Debug traits. That seems to warrant some discussion.