prometheus_exporter_base -> prometheus
Due to the former being unmaintained and vulnerable, switch the exporter
library.

I opted for this library and not the official prometheus-client crate
because I had trouble figuring out a way to deal with globally available
metrics and didn't want to rewrite the entire thing to thread the state
through faythe.

The whole thing is basically a slight modification to the upstream hyper
example at:

https://github.com/tikv/rust-prometheus/blob/1d3174bf5ddf056dcb0fe59e06cad4ef42ebec68/examples/example_hyper.rs